requirements.txt
Flask
flask_cors
bcrypt
python-dotenv
opencv-python 